Refactoring: Moved InstantCoverter to package common/persistence
