+++ backend/src/main/java/org/educama/shipment/model/Invoice.java
@@ -1,5 +1,6 @@
 package org.educama.shipment.model;
 
+import org.educama.common.persistence.InstantConverter;
 import org.springframework.data.jpa.domain.AbstractPersistable;
 
 import javax.persistence.*;
